requirements.txt
# Core dependencies
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0
transformers>=4.30.0
accelerate>=0.20.0

# Vector database and retrieval
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU acceleration

# Language model APIs
openai>=1.0.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
PIL>=8.0.0

# Configuration and utilities
omegaconf>=2.3.0
tqdm>=4.64.0
nest-asyncio>=1.5.0

# Multimodal models
flmr>=0.1.0  # Fine-grained Late-interaction Multi-modal Retriever

# Evaluation
datasets>=2.10.0
scikit-learn>=1.0.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0

# Additional utilities
requests>=2.28.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
